Updated relative paths, moved old isolated scripts into a subdir, and removed Laurel from test email params
