+++ bin/mail/common/mailing-list-utils.rb
