+++ bin/mail/common/split
