+++ bin/mail/common/test-params-all.csv
@@ -1,5 +1,3 @@
 email,name,international
-laurel+international@code.org,Laurel International,true
 andrew+international@code.org,Andrew International,true
-laurel+us@code.org,Laurel US,false
 andrew+us@code.org,Andrew US,false
