+++ bin/mail/common/test-params-international.csv
@@ -1,3 +1,2 @@
 email,name,international
-laurel+international@code.org,Laurel International,true
 andrew+international@code.org,Andrew International,true
