+++ bin/mail/common/test-params-us.csv
@@ -1,3 +1,2 @@
 email,name,international
-laurel+us@code.org,Laurel US,false
 andrew+us@code.org,Andrew US,false
