+++ bin/mail/hoc-2015-dec-6-mailing/generate-all-mailing-lists.rb
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
-require_relative '../mailing-common/mailing-list-utils'
+require_relative '../common/mailing-list-utils'
 
 # 2015 organizers (kind_s:HocSignup2015)
 generate('hoc-organizers')
