+++ bin/mail/hoc-2015-dec-6-mailing/send-all-test-mails
@@ -1,4 +1,4 @@
 #!/bin/sh
 
-./send-to-mailing-list 12-07-hoc-general mailing-common/test-params-all.csv
+./send-to-mailing-list 12-07-hoc-general common/test-params-all.csv
