+++ bin/mail/hoc-2015-dec-6-mailing/send-organizers-test-mails
@@ -1,4 +1,4 @@
 #!/bin/sh
 
-./send-to-mailing-list 12-07-hoc-organizers mailing-common/test-params-us.csv
+./send-to-mailing-list 12-07-hoc-organizers common/test-params-us.csv
