+++ bin/mail/hoc-2015-nov-19-mailing/send-test-mails
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-./send-to-mailing-list 11-19-teachers mailing-common/test-params-us.csv
+./send-to-mailing-list 11-19-teachers common/test-params-us.csv
