+++ bin/mail/hoc-2015-nov-30-mailing/generate-hoc-organizers
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
-require_relative '../mailing-common/mailing-list-utils'
+require_relative '../common/mailing-list-utils'
 
 query = <<EOS
 kind_s:HocSignup2015
