+++ bin/mail/hoc-2015-nov-30-mailing/send-test-mails
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-./send-to-mailing-list 11-30-hoc-organizers mailing-common/test-params-all.csv
+./send-to-mailing-list 11-30-hoc-organizers common/test-params-all.csv
