+++ bin/mail/hoc-2015-nov-6-mailing/generate-hoc-teachers
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
-require_relative '../mailing-common/mailing-list-utils'
+require_relative '../common/mailing-list-utils'
 
 # Teachers requesting volunteers
 teacher_query = <<EOS
