+++ bin/mail/hoc-2015-nov-6-mailing/send-hoc-teachers-test-mails
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-./send-to-mailing-list 11-3-teachers mailing-common/test-params-all.csv
+./send-to-mailing-list 11-3-teachers common/test-params-all.csv
