+++ bin/mail/hoc-2015-nov-6-mailing/send-hoc-volunteers-test-mails
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-./send-to-mailing-list 11-3-volunteers mailing-common/test-params-us.csv
+./send-to-mailing-list 11-3-volunteers common/test-params-us.csv
