+++ bin/mail/hoc-2015-nov-9-mailing/generate-2015-organizers
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
-require_relative '../mailing-common/mailing-list-utils'
+require_relative '../common/mailing-list-utils'
 
 query = <<EOS
 kind_s:HocSignup2015
