+++ bin/mail/hoc-2015-nov-9-mailing/send-2015-organizers-test-mails
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-./send-to-mailing-list 11-9-hoc-organizers mailing-common/test-params-all.csv
+./send-to-mailing-list 11-9-hoc-organizers common/test-params-all.csv
