+++ bin/mail/hoc-2015-nov-9-mailing/send-general-final-test-mails
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-./send-to-mailing-list 11-9-hoc-general mailing-common/test-params-all.csv
+./send-to-mailing-list 11-9-hoc-general common/test-params-all.csv
