+++ bin/mail/hoc-2015-oct-26-mailing/generate-hoc-engineers
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
-require_relative '../mailing-common/mailing-list-utils'
+require_relative '../common/mailing-list-utils'
 
 volunteer_query = <<EOS
 kind_s:VolunteerEngineerSubmission2015
