+++ bin/mail/hoc-2015-oct-26-mailing/generate-uk-poster-list
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
-require_relative '../mailing-common/mailing-list-utils'
+require_relative '../common/mailing-list-utils'
 
 query = <<EOS
 kind_s:"HocSignup2015" && location_country_code_s: "GB"
