+++ bin/mail/hoc-2015-oct-26-mailing/send-hoc-engineers-test-mails
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-./send-to-mailing-list 10-28-hoc-engineers mailing-common/test-params-all.csv
+./send-to-mailing-list 10-28-hoc-engineers common/test-params-all.csv
