+++ bin/mail/hoc-2015-survey/generate-organizer-list
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
-require_relative '../mailing-common/mailing-list-utils'
+require_relative '../common/mailing-list-utils'
 require 'set'
 
 query = <<EOS
