+++ bin/mail/hoc-2015-survey/send-hoc-volunteers-test-mails
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-./send-to-mailing-list 2015-12-17-volunteers mailing-common/test-params-us.csv
+./send-to-mailing-list 2015-12-17-volunteers common/test-params-us.csv
