+++ bin/mail/mailing-1-29-white-house/generate-all-mailing-lists.rb
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
-require_relative '../mailing-common/mailing-list-utils'
+require_relative '../common/mailing-list-utils'
 
 # ALL U.S. code studio + HOC teachers
 generate('us-teacher')
