+++ bin/mail/mailing-1-5-advocacy-ca/send-ca-test-mails
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-./send-to-mailing-list 1-5-advocacy-ca mailing-common/test-params-us.csv
+./send-to-mailing-list 1-5-advocacy-ca common/test-params-us.csv
