+++ bin/mail/mailing-2-8-teacher-recruitment/send-test-mails
@@ -7,7 +7,7 @@ unless template_name
   exit
 end
 
-template_file = File.join('../../pegasus/emails', "#{template_name}.md")
+template_file = File.join('../../../pegasus/emails', "#{template_name}.md")
 raise "Missing email template #{template_file}" unless File.exist? template_file
 
 `../send-to-mailing-list #{template_name} test-params.csv`
