+++ bin/mail/old-isolated-scripts/generate-mailing-list-everybody
@@ -1,13 +1,7 @@
 #!/usr/bin/env ruby
-require_relative '../../pegasus/src/env'
+require_relative '../../../pegasus/src/env'
 require 'cdo/solr'
 require src_dir 'database'
-
-# Generates a list of teachers based on these filters:
-# Petition parents, software engineers, none of the above - US ONLY
-# Volunteer engineer list - US ONLY
-# HocSignup2014- US ONLY
-# CSEdWeekEvent2013- US ONLY
 
 SOLR = Solr::Server.new(host: CDO.solr_server)
 
@@ -44,17 +38,12 @@ UNSUBSCRIBERS = {}.tap do |results|
 end
 puts "#{UNSUBSCRIBERS.count} unsubscribers loaded."
 
-TEACHERS = {}.tap do |results|
-  (
-  query_contacts(q: 'kind_s:"Petition" && create_ip_country_s:"United States"', fq: '-role_s:"student" && -role_s:"educator" && -role_s:"district_admin" && -role_s:"principal" && -role_s:"superintendent" && -role_s:"teacher"') +
-      query_contacts(q: 'kind_s:"VolunteerEngineerSubmission" && create_ip_country_s:"United States"') +
-      query_contacts(q: 'kind_s:"HocSignup2014" && create_ip_country_s:"United States"') +
-      query_contacts(q: 'kind_s:"CSEdWeekEvent2013" && create_ip_country_s:"United States"')
-  ).each do |i|
+EVERYBODY = {}.tap do |results|
+  query_contacts(q: '-email_s:""').each do |i|
     email = i[:email].downcase.strip
     results[email] = i unless UNSUBSCRIBERS[email]
   end
 end
-puts "#{TEACHERS.count} teachers loaded."
+puts "#{EVERYBODY.count} recipients loaded."
 
-export_contacts_to_csv TEACHERS, "us-teachers.csv"
+export_contacts_to_csv EVERYBODY, 'everybody.csv'
