+++ bin/mail/old-isolated-scripts/generate-mailing-list-from-query
@@ -1,7 +1,10 @@
 #!/usr/bin/env ruby
-require_relative '../../pegasus/src/env'
+require_relative '../../../pegasus/src/env'
 require 'cdo/solr'
 require src_dir 'database'
+
+raise "Usage: $0 'solr query'" unless ARGV.length == 1
+solr_query = ARGV[0]
 
 SOLR = Solr::Server.new(host: CDO.solr_server)
 
@@ -20,24 +23,15 @@ def export_contacts_to_csv(contacts, path)
 end
 
 def query_contacts(params)
+  fields = []
   fields = params[:fields] if params[:fields]
 
   [].tap do |results|
-    SOLR.query(params.merge(rows: 10000)).each do |i|
+    SOLR.query(params.merge(rows: 10000, fl: ['email_s', 'name_s'] + fields)).each do |i|
       i = yield(i) if block_given?
       results << {email: i['email_s'].downcase.strip, name: i['name_s']}.merge(i.slice(*fields)) if i
     end
   end
-end
-
-def query_contacts_with_location(params)
-  raise "there must be a query" if params[:q].nil?
-
-  params = params.dup
-  params[:q] += ' && ' + location_query if location_query
-  params[:fq] = location_filter_query
-
-  query_contacts(params)
 end
 
 UNSUBSCRIBERS = {}.tap do |results|
@@ -48,14 +42,12 @@ UNSUBSCRIBERS = {}.tap do |results|
 end
 puts "#{UNSUBSCRIBERS.count} unsubscribers loaded."
 
-TEACHERS = {}.tap do |results|
-  (
-  query_contacts(q: 'kind_s:"HocSignup2014" && location_country_s:"Brazil"')
-  ).each do |i|
+recipients = {}.tap do |results|
+  query_contacts(q: solr_query).each do |i|
     email = i[:email].downcase.strip
     results[email] = i unless UNSUBSCRIBERS[email]
   end
 end
-puts "#{TEACHERS.count} teachers loaded."
+puts "#{recipients.count} recipients > mailing-list.csv"
 
-export_contacts_to_csv TEACHERS, "brazil-teachers.csv"
+export_contacts_to_csv recipients, 'mailing-list.csv'
