+++ bin/mail/old-isolated-scripts/generate_professional_development_workshop_teachers_report_csv
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
-require File.expand_path('../../pegasus/src/env', __FILE__)
+require File.expand_path('../../../pegasus/src/env', __FILE__)
 require src_dir 'database'
 require pegasus_dir 'helpers/professional_development_workshop_helpers'
 require 'cdo/csv'
